+++ src/main/java/com/apajac/acolhimento/services/interfaces/AtualizarStatusUsuarioService.java
@@ -1,5 +1,5 @@
 package com.apajac.acolhimento.services.interfaces;
 
 public interface AtualizarStatusUsuarioService {
-    void updateStatusUsuario(Integer id, Boolean status);
+    void updateStatusUsuario(Long id, Boolean status);
 }
